backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
azure-identity
azure-cosmos
openai>=1.45.0
